% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blanket_statsments.R
\name{blanket_c_statistic}
\alias{blanket_c_statistic}
\title{Generic wrapper method to calculate C-statistics}
\usage{
blanket_c_statistic(df, model_list, modality = "logistic", verbose = FALSE)
}
\arguments{
\item{df}{data.frame containing the data set. If evaluating independently, use the test set.}

\item{model_list}{list of statistical models of type lm, glm or coxph to be evaluated.}

\item{modality}{character specifying model type. Currently accepts 'linear', 'logistic', and 'cox'}

\item{verbose}{logical. TRUE activates printout messages.}
}
\value{
list of doubles with the AUC values for the evaluated models on the specified data set.
}
\description{
Calculate concordance statistics for a list of statistical models on the same data set
}
\author{
J. Peter Marquardt
}
