% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blanket_statsments.R
\name{.build_model_formula}
\alias{.build_model_formula}
\title{Build formula for statistical models}
\source{
\link[basecamb]{build_model_formula}
}
\usage{
.build_model_formula(outcome, predictors, censor_event = NULL)
}
\arguments{
\item{outcome}{character denoting the column with the outcome.}

\item{predictors}{vector of characters denoting the columns with the
predictors.}

\item{censor_event}{character denoting the column with the censoring event,
for use in Survival-type models.}
}
\value{
formula for use in statistical models
}
\description{
Build formula used in statistical models from vectors of strings. Copied from basecamb package to avoid dependency
}
\author{
J. Peter Marquardt
}
