% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blanket_statsments.R
\name{table_blanket_redundancies}
\alias{table_blanket_redundancies}
\title{Table results of blanket redundancy analysis}
\usage{
table_blanket_redundancies(blanket_redundancies, digits = 2)
}
\arguments{
\item{blanket_redundancies}{list of lists of redun objects generated by blanket_redundancy_analysis()}

\item{digits}{integer number of decimals to include}
}
\value{
a data.frame tabling the key results
}
\description{
Table results of a blanket redundancy analysis on a list of existing models
}
\seealso{
[table_predictors()], [blanket_redundancy_analysis()]
}
\author{
J. Peter Marquardt
}
