% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blanket_statsments.R
\name{blanket_redundancy_analysis}
\alias{blanket_redundancy_analysis}
\title{Blanket redundancy analysis}
\usage{
blanket_redundancy_analysis(
  model_list,
  r2_threshold = 0.9,
  nk = 0,
  verbose = FALSE
)
}
\arguments{
\item{model_list}{a list of statistical regression model of class linear, logistic or coxph}

\item{r2_threshold}{float threshold value to consider a parameter redundant}

\item{nk}{number of knots in splicing}

\item{verbose}{ctivate printouts of key findings}
}
\value{
an list of objects of class "redun"
}
\description{
Perform a blanket redundancy analysis on a list of existing models
}
\seealso{
[blanket_stats()]
}
\author{
J. Peter Marquardt
}
