% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blanket_statsments.R
\name{redundancy_analysis}
\alias{redundancy_analysis}
\title{Redundancy analysis}
\usage{
redundancy_analysis(model, r2_threshold = 0.9, nk = 0)
}
\arguments{
\item{model}{a statistical regression model of class linear, logistic or coxph}

\item{r2_threshold}{float threshold value to consider a parameter redundant}

\item{nk}{number of knots in splicing}
}
\value{
an object of class "redun"
}
\description{
Perform a redundancy analysis on an existing model
}
\author{
J. Peter Marquardt
}
